state: document and implement gameplay order
Using a shuffling scheme as attempted previously has some shortcomings:

1. Asymmetry in behavior for player 1 x others
2. If messages include nonces, player 1 can identify encrypted values
   even after they are encrypted by other players, giving an unfair
   advantage.
3. If messages do not include nonces, then a single nonce must be used
   with the same key to encrypt multiple values, what makes it a weaker
   scheme.

Note that it is incorrect to encrypt a []byte with one call to Encrypt
and then shuffle the bytes, as it won't be possible to decode the
shuffled bytes with the same key.

The new scheme is symmetric, and gives no advantage to any player.